@@ -4,8 +4,8 @@
             <strong>{{ config('app.name', 'Laravel') }}</strong> by <a href="https://www.maximejund.fr">Maxime
                 Jund</a>. The source code is
             licensed
-            <a href="http://opensource.org/licenses/mit-license.php">MIT</a>. The website content
-            is licensed <a href="http://creativecommons.org/licenses/by-nc-sa/4.0/">CC BY NC SA 4.0</a>.
+            <a href="https://opensource.org/licenses/mit-license.php">MIT</a>. The website content
+            is licensed <a href="https://creativecommons.org/licenses/by-nc-sa/4.0/">CC BY NC SA 4.0</a>.
             <br />Made with <span class="icon-text"><span class="icon"><i
                         class="fa-solid fa-heart has-text-danger"></i></span></span>for <a
                 href="https://www.crechelesfripouilles.com/">Crèche parentale
